[#24] recoil atom 파일 분리
